[feature/22] (disgusting) version of on-demand loading of languages.
This commit adds support on-demand loading of languages, which runs KTS with no
tree-sitter data at first, and will only load the ones that are asked by buffer.

Most of the design is fine, but highlighting/injection is _terrible_, because of
the design of the `highlight()` and injection callback of
`tree-sitter-highlight`. There is no way around it:

- `Rc<RefCell<Language>>` will not allow to obtain a `Option<&Language>` by
  leaving the `FnMut` as the rc will be living in the closure stackframe.
- Using a second map only for injected language and returning references to
  those freshly inserted items won’t work either, as those will be captured
  mutably and we cannot escape a reference (even immutable) from them. It’s
  basically the `&mut T -> &Q` mutable borrow lock, but with `FnMut` in between.
- I swear I tried other ideas which all failed and I just want to get done with
  that shit.

May you show mercy on that shitty code.

References: https://todo.sr.ht/~hadronized/kak-tree-sitter-features/22