Reimplement proper async IO decoupling.
That commit moves fifos to the IO handler and introduces commands for the handler to read from. This commit
also introduces back buffer, a mechanism with which the handler can copy contents of strings sent by the IO
handler and send back strings to the IO handler to prevent allocating like a crazy cat on steroids.

References: https://todo.sr.ht/~hadronized/kak-tree-sitter-bugs/30